requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
SQLAlchemy==2.0.43
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.16.5
greenlet==3.2.4
openai==1.107.1
pgvector==0.4.2
tiktoken==0.12.0
trafilatura==2.0.0
httpx==0.27.2
